Move ColorAdapter into serialization package
